+++ src/transaction/firestore-pubsub/readonly-state-transaction.ts
@@ -4,14 +4,8 @@ import type {
 } from '@causa/runtime';
 import type { Type } from '@nestjs/common';
 import { Transaction } from 'firebase-admin/firestore';
-import {
-  getReferenceForFirestoreDocument,
-  makeFirestoreDataConverter,
-} from '../../firestore/index.js';
-import {
-  getSoftDeletedFirestoreCollectionMetadataForType,
-  type SoftDeletedFirestoreCollectionMetadata,
-} from './soft-deleted-collection.decorator.js';
+import { getReferenceForFirestoreDocument } from '../../firestore/index.js';
+import { getSoftDeleteInfo } from './soft-deleted-collection.decorator.js';
 import type { FirestoreCollectionResolver } from './types.js';
 
 /**
@@ -19,19 +13,6 @@ import type { FirestoreCollectionResolver } from './types.js';
  */
 export type FirestoreReadOnlyStateTransactionOption =
   ReadOnlyTransactionOption<FirestoreReadOnlyStateTransaction>;
-
-/**
- * Information about soft-deletion for a given document.
- */
-export type SoftDeleteInfo<T extends object> = Omit<
-  SoftDeletedFirestoreCollectionMetadata,
-  'deletedDocumentsCollectionSuffix'
-> & {
-  /**
-   * The reference to the soft-deleted document.
-   */
-  ref: FirebaseFirestore.DocumentReference<T>;
-};
 
 /**
  * A {@link ReadOnlyStateTransaction} that uses Firestore for state storage.
@@ -57,34 +38,6 @@ export class FirestoreReadOnlyStateTransaction
     readonly collectionResolver: FirestoreCollectionResolver,
   ) {}
 
-  /**
-   * Returns the soft-delete information for a given document, if the document type supports soft-deletion.
-   *
-   * @param activeDocRef The reference to the active document.
-   * @param type The type of the document.
-   * @returns The soft-delete information for the document, or `null` if the document type does not support
-   *   soft-deletion.
-   */
-  protected getSoftDeleteInfo<T extends object>(
-    activeDocRef: FirebaseFirestore.DocumentReference<T>,
-    type: Type<T>,
-  ): SoftDeleteInfo<T> | null {
-    const softDeleteMetadata =
-      getSoftDeletedFirestoreCollectionMetadataForType(type);
-    if (!softDeleteMetadata) {
-      return null;
-    }
-
-    const { deletedDocumentsCollectionSuffix: suffix, ...info } =
-      softDeleteMetadata;
-    const softDeleteCollection = activeDocRef.firestore
-      .collection(`${activeDocRef.parent.path}${suffix}`)
-      .withConverter(makeFirestoreDataConverter(type));
-    const ref = softDeleteCollection.doc(activeDocRef.id);
-
-    return { ...info, ref };
-  }
-
   async get<T extends object>(
     type: Type<T>,
     entity: Partial<T>,
@@ -103,7 +56,7 @@ export class FirestoreReadOnlyStateTransaction
       return activeDocument;
     }
 
-    const softDeleteInfo = this.getSoftDeleteInfo(activeDocRef, type);
+    const softDeleteInfo = getSoftDeleteInfo(activeDocRef, type);
     if (!softDeleteInfo) {
       return null;
     }